moved admin info button
